requirements.txt
streamlit==1.19.0
streamlit-authenticator==0.2.1
PyYAML==5.4.1
pandas==1.2.4
matplotlib==3.3.4
japanize-matplotlib==1.1.3